+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+- Opensource release
+
 ## 0.1.3 - 2018-06-02
 ### Fixed
 - Add `axi_lite_to_reg.sv` to list of source files.