+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - AXI to reg interface
 - Opensource release
 
+### Changed
+- Updated AXI dependency
+
 ## 0.1.3 - 2018-06-02
 ### Fixed
 - Add `axi_lite_to_reg.sv` to list of source files.