+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## 0.2.0 - 2020-12-30
 ### Fixed
 - Fix bug in AXI-Lite to register interface conversion
 - Fix minor style problems (`verible-lint`)